combined irc.json and animes.json

+++ horriblesubs.py
@@ -17,19 +17,19 @@ from configloader.ircsettingsloader import IrcSettingsLoader
 from lib.packagehelper import get_diff_episodes, get_episode_package
 from lib.tcpdownloader import tcpdownload, tcpdownload_one
 from lib.xdccparser import parse_name, search
+from lib.configloader import configloader
 
 logging.basicConfig(level=logging.INFO)
 
 
 BASEURL = "http://horriblesubs.info"
-ISL = IrcSettingsLoader()
-ASL = AnimeSettingsLoader()
+CONFIG = configloader()
 
-
-def get_local_episodes(name):
+def get_local_episodes(anime_folder, name):
     """return a list of files of a anime-folder inside ANIME_FOLDER"""
+    
     episodes = []
-    path = os.path.join(ISL.get_anime_folder(), name)
+    path = os.path.join(anime_folder, name)
     if not os.path.isdir(path):
         os.makedirs(path)
         return episodes
@@ -42,13 +42,13 @@ def get_local_episodes(name):
     return episodes
 
 
-def delete_local_episodes(anime, episode):
+def delete_local_episodes(anime_folder, anime, episode):
     """delete the local ep"""
     filename = "[HorribleSubs] " + anime + " - " + episode + " [720p].mkv"
-    path = os.path.join(ISL.get_anime_folder(), anime, filename)
+    path = os.path.join(anime_folder, anime, filename)
     if os.path.isfile(path):
-        print("REMOVE: '%s'" % path)
-        os.remove(path)
+        y = input("REMOVE? (y/n): '{}'".format(path))
+        if y.lower() == "y": os.remove(path)
 
 
 def print_json(data):
@@ -57,20 +57,20 @@ def print_json(data):
     print(json.dumps(parsed, indent=4, sort_keys=True))
 
 
-def check_animes(animes):
+def check_animes(animes, irc_config):
     """compare with caching"""
     result = []
     table_data = []
     for idx, show in enumerate(animes):
-        packages = search(show, ISL.get_default_res())
-        local = get_local_episodes(show)
+        packages = search(show, irc_config["default_res"])
+        local = get_local_episodes(irc_config["anime_folder"], show)
         diff = get_diff_episodes(packages, local)
         if diff:
             table_data.append([idx + 1, show, colored(str(diff), 'red')])
             for episode in diff:
-                delete_local_episodes(show, episode)
+                delete_local_episodes(irc_config["anime_folder"], show, episode)
                 package = get_episode_package(
-                    packages, episode, ISL.get_default_bot())
+                    packages, episode, irc_config["default_bot"])
                 result.append(package)
         else:
             table_data.append(
@@ -94,18 +94,18 @@ def check_irc(irc_queue_in, irc_queue_out):
     irc_queue_in.put("clear")
 
 
-def boot_up():
+def boot_up(anime_folder):
     """boot up. ASL ISL anime folder"""
     cls()
-    if ASL.is_loaded() and ISL.is_loaded():
+    if CONFIG.is_loaded():
         print(
             colored("<] configs successfull loaded [>\n", "green").center(80))
     else:
         print(
             colored(">[ configs failed to load. Exit ]<\n", "red").center(80))
         sys.exit(1)
-    if not os.path.isdir(ISL.get_anime_folder()):
-        print(colored("anime folder not found: {}".format(ISL.get_anime_folder()), "red"))
+    if not os.path.isdir(anime_folder):
+        print(colored("anime folder not found: {}".format(anime_folder), "red"))
         sys.exit(1)
 
 
@@ -114,7 +114,7 @@ def setup_irc():
     manager = mp.Manager()
     irc_queue_in = manager.Queue()
     irc_queue_out = manager.Queue()
-    irc = IrcLib(ISL, irc_queue_in, irc_queue_out)
+    irc = IrcLib(CONFIG.get_irc(), irc_queue_in, irc_queue_out)
     irc.start()
     check_irc(irc_queue_in, irc_queue_out)
     return irc_queue_in, irc_queue_out
@@ -123,17 +123,18 @@ def setup_irc():
 def main():
     """main"""
     colorama.init()
-    boot_up()
-    animes = ASL.get_watching()
+    boot_up(CONFIG.get_irc()["anime_folder"])
+    animes = CONFIG.get_animes()["watching"]
+   
     if len(animes) <= 0:
         print("no subscribed animes. 'python sub_gui.py' for gui or animes.json")
-    data = check_animes(animes)
+    data = check_animes(animes, CONFIG.get_irc())
     if not data:
         print(colored("<] nothing to do. [>\n", "green").center(80))
     else:
         irc_queue_in, irc_queue_out = setup_irc()
         if input("press enter to start downloading...") == "":
-            if ISL.get_multiprocessing() == 0:
+            if CONFIG.get_irc()["multiprocessing"] == 0:
                 tcpdownload_one(irc_queue_in, irc_queue_out, data)
             else:
                 tcpdownload(irc_queue_in, irc_queue_out, data)
